+++ docker/install-homebrew-deps.bash
@@ -4,8 +4,8 @@ set -euxo pipefail
 
 export DEBIAN_FRONTEND=noninteractive
 
-apt-get update
-apt-get install -y curl git jq make pass sudo wget
+(yes || true) | unminimize
+apt-get install -y build-essential curl file git make man-db procps sudo wget
 
 rm -rf /var/lib/apt/lists/*
 rm "$0"