+++ packaging/deb-install.bash
@@ -2,4 +2,8 @@
 
 set -euxo pipefail
 
+if [[ -z "$(ls -A /var/lib/apt/lists)" ]]; then
+    sudo apt update
+fi
+
 sudo apt reinstall "./out/${VERSION}/pass-extension-ln-${VERSION}.deb"